Add support for versions without macros
Fixes #17

This is slower so it should only be used on the older versions,
but technically it's compatible with newer versions too. I'm not doing
any feature detection for it so it's in a separate branch and will live
here. There is no guarantee that features will be ported to this version